pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "musicleague"
version = "0.1.0"
description = "musicleague project"
requires-python = ">=3.13"
dependencies = [
	"pandas>=2.2",
	"python-dotenv>=1.0",
	"spotipy>=2.24",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
